Added temp reading code, updated uploading code, untested

+++ src/upload.h
@@ -1,1 +1,1 @@
-int UploadTemp(float temp);
+int UploadTemp(double temp);
